requirements.txt
websocket-client
requests
click
datetime
tqdm
podman-compose
openai
pandas
matplotlib
seaborn
tabulate